% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CDRexamples}
\alias{CDRexamples}
\title{Example data to test functions.}
\format{A data frame with 23 rows and 8 variables:
\describe{
  \item{id}{id variable}
  \item{test}{example described}
  \item{glo5}{memory}
  \item{glo6}{orientation}
  \item{glo7}{judgement, problemsolving}
  \item{glo8}{community affairs}
  \item{glo9}{home and hobbies}
  \item{glo10}{Personal care}
  \item{NACC_alg}{scoring from the official website}
}}
\usage{
CDRexamples
}
\description{
A data set that contains different coding situations and how the algoritm
handles this.
}
\keyword{datasets}
