% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_CDR_user.R
\name{code_CDR}
\alias{code_CDR}
\title{Assignment of CDR rating}
\usage{
code_CDR(df, id_name = "lopnr", deb = FALSE)
}
\arguments{
\item{df}{a data frame}

\item{id_name}{the name of the primary ID variable in the data set}

\item{deb}{debug yes or no (will slow function down)}
}
\value{
a data frame
}
\description{
code_CDR adds a variable called "CDR" to a data set according to the scoring
algorithm by Morris (1993).
}
\examples{
code_CDR(df = CDRexamples, id_name = "id", deb = FALSE)
code_CDR(df = CDRexamples, id_name = "id", deb = TRUE)
}
